+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Removed Python 3.5 support (#404)
 - Updated list of base API resources to include `aliases`, `git_repos`, `json_values`, `services`, and `storage_hosts` so that they show up in the sphinx docs (#406)
 - Update the API spec at `civis/tests/civis_api_spec.json` so that new endpoints are included (e.g., `/exports/files/csv`) (#407)
+- Refactor file cleaning logic for `civis.io.civis_file_to_table` (#405)
 
 ### Fixed
 - Fixed a workflows usage example in `docs/source/client.rst` that had an incorrect endpoint. (#409)